refactor: split startup dialogs
